1、修改文件传输的端口 2、梳理报错

+++ server/src/main/java/tran/server/io/netty/model/ClientHeartbeatAckPackage.java
@@ -1,20 +1,15 @@
-//
-// Source code recreated from a .class file by IntelliJ IDEA
-// (powered by Fernflower decompiler)
-//
-
 package tran.server.io.netty.model;
 
-import com.aww.fts.socket.netty.decoder.FTProtocolDecoder;
 import io.netty.buffer.ByteBuf;
 import io.netty.buffer.Unpooled;
+import tran.server.io.netty.decoder.FTProtocolDecoder;
 
 public class ClientHeartbeatAckPackage extends FTPackage {
     private int clientId;
     private long syncId;
 
     public ClientHeartbeatAckPackage() {
-        this.setPackageType((short) FTProtocolDecoder.PACKAGE_TYPE_CLIENT_HEARTBEAT_SYNC_ACK);
+        this.setPackageType(FTProtocolDecoder.PACKAGE_TYPE_CLIENT_HEARTBEAT_SYNC_ACK);
     }
 
     public ByteBuf getByteBuf() {
